pyproject.toml
[project]
name = "agent-loop-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "black>=25.1.0",
    "fastmcp>=2.11.3",
    "ipykernel>=6.30.1",
    "mlflow>=3.3.0",
    "openai>=1.100.2",
    "openai-agents[viz]>=0.2.8",
    "python-docx>=1.2.0",
    "python-dotenv>=1.1.1",
]
